Started creating barometer class
